# ARROW-9605: [C++] Speed up aggregate min/max compute kernels on integer types