1. Use BitBlockCounter to speedup the performance for typical 0.01% null probability data.
2. Enable compiler auto SIMD vectorize for no-nulls on int types. Float/Double use fmin/fmax to handle NaN which can't be vectorize by compiler.
3. Also add test case to cover different null probability.