@jianxind 2020-08-05
Need a rebase after https://github.com/apache/arrow/pull/7903