@jianxind 2020-08-12
Ping. @wesm @pitrou 

Could you help to review this? Similar approach to sum kernel, use compiler to vectorise the NoNulls part, use BitBlockCounter on the 0.01% data. https://github.com/apache/arrow/pull/7870 add the benchmark item for MinMax kernel.

Thanks.