@jianxind 2020-08-13
@ursabot benchmark --suite-filter=arrow-compute-aggregate-benchmark --benchmark_filter="MinMax"