@jianxind 2020-09-02
> ARROW_USER_SIMD_LEVEL=none

Below is the cmd I used, and compiler vectorise happens only on Int types.
```
ARROW_USER_SIMD_LEVEL=avx2 ./release/arrow-compute-aggregate-benchmark --benchmark_filter=MinMaxKernelInt64
ARROW_USER_SIMD_LEVEL=none ./release/arrow-compute-aggregate-benchmark --benchmark_filter=MinMaxKernelInt64
```