@pitrou 2020-09-02
Ah, I also had {{-DARROW_SIMD_LEVEL=AVX2}} in CMake. Without it I do see a difference.