@pitrou 2020-09-02
Rebased.