@pitrou 2020-09-02
Test failures are unrelated, will merge.