@felipecrv 2024-08-14
> Ah, I also had `-DARROW_SIMD_LEVEL=AVX2` in CMake. Without it I do see a difference.

I've came here after looking at the code and being confused. It sounds like there was never a need to instantiate templates with a `SimdLevel` parameter given that the SIMD comes from compiler auto-vectorization instead of the kernel code doing something different.

I might be wrong, in that case, I would love to have a pointer to the specialized code.