@pitrou 2024-08-15
@felipecrv I may be misunderstanding your question, but the `SimdLevel` template parameter is used as a disambiguation against ODR issues (you would get multiple times the same function with a different generated code due to different compiler options).