@felipecrv 2024-08-15
> @felipecrv I may be misunderstanding your question, but the `SimdLevel` template parameter is used as a disambiguation against ODR issues (you would get multiple times the same function with a different generated code due to different compiler options).

I understand that, but if no specialization exists for different SIMD levels, we don't need more than the `SimdLevel::NONE` variation of the kernel. There is no need to populate the kernel array for `SimdLevel::AVX2` or `AVX512` if we are not performing any sort of runtime dispatching based on runtime CPU capability checks.